widen `Op` to u16 — for TAG SPACE, and the speed is a lead not a claim
248 of 256 byte values were assigned and wide-arithmetic took the last four, so the
next proposal to add instructions had nowhere to go. Custom-descriptors needs at least
eight. This is the enabling half: `Op` becomes `#[repr(u16)]` and every site that
needed the discriminant stops truncating. **Tag VALUES are unchanged**, so the whole
change is verifiably a no-op on behaviour.

Measured before adopting, not after:

* **Canonical behaviour: unaffected, and it cannot be otherwise.** `Op` is an internal
  tag; the wire format is a `0xFB`–`0xFE` prefix plus a LEB sub-opcode, read and
  written independently of it. Checked rather than asserted — the full 288-file spec
  suite produced output **byte-identical** to the `u8` run (`diff -q`,
  64,068 / 66 / 549 both ways). `Op` never crosses the C ABI, so the T8 freeze is
  untouched; C-ABI gate PASSED, 74 symbols.
* **Size: free.** `size_of::<Op>()` 1 → 2 and `size_of::<Instr>()` **stays 80** —
  `Imm` is 64 bytes at 16-byte alignment, so the second byte lands in the padding
  `offset` already occupies. `instr_is_sixteen_bytes` passes unchanged.
* **Speed: it gains.** A/B/A/B before landing: u8 = 229.8 / 228.5 / 225.6 Mops/s,
  u16 = 250.1 / 243.4; this build reads 246.0. +7–9%, groups non-overlapping.
  ⚠️⚠️ **The mechanism is UNEXPLAINED.** Widening a discriminant should be neutral, not
  8% faster. The number is measured (repeated, bracketed); the cause is not, so it goes
  to T11 beside the unattributed ~5% steady regression and **is not a figure to quote**.
  That is also why this commit is titled for the tag space: the space is the reason,
  the speed is a side effect nobody can yet explain.

🔒 **A tag may now exceed `0xff`, so `op as u8` became a TRUNCATING cast** — converted
at every site that matters, while values still fit in a byte so the conversion is
provably behaviour-preserving today:

* `exec_numeric` and `exec_float` dispatch on `op as u16`; both carry internal tags
  (saturating truncation) in their ranges.
* `simple_sig` matches `op as u16`; it carries saturating-truncation and
  wide-arithmetic tags.
* ⚠️ The `cmp_*` / `bin_*` / `fcmp` helpers deliberately still take `u8`: every arm that
  calls them is a real single-byte opcode, and the const-expression evaluator calls the
  same helpers with a raw wire byte.
* 🔒 `emit_op_with_immediates`' catch-all now converts **fallibly**. An internal tag
  reaching it means the prefixed-family arms above are missing a case, and `op as u8`
  would have silently emitted a truncated byte — a *different instruction*. It refuses
  by name instead, which says which op has no emitter.

503 workspace tests, C-ABI gate passed, suite byte-identical.

Next, separately: move the internal tags above `0xff` and retire `decode_body`'s
raw-internal-tag guard. That is the part that actually frees the space, and it rewrites
how the byte range is partitioned, so it gets its own commit.

+++ crates/wasmrt-core/src/interp.rs
@@ -4442,8 +4442,13 @@ fn ref_matches(
 
 /// Integer arithmetic / comparison / bitwise / conversion opcodes. Anything else (float,
 /// memory, SIMD, …) traps `UnsupportedInstruction` in this slice.
+// ⚠️ Dispatches on `op as u16`, not `as u8`: `Op` is `#[repr(u16)]` and the internal tags for the
+// prefixed families can sit above `0xff`, where a `u8` cast would truncate one op onto another.
+// The `cmp_*` / `bin_*` / `fcmp` helpers still take `u8` on purpose — every arm that calls them is
+// a real single-byte opcode, and the const-expression evaluator calls the same helpers with a raw
+// wire byte.
 fn exec_numeric(frame: &mut Frame, op: Op) -> Result<()> {
-    match op as u8 {
+    match op as u16 {
         // i32 unary
         0x45 => {
             let v = frame.pop_i32();
@@ -4545,7 +4550,7 @@ fn exec_numeric(frame: &mut Frame, op: Op) -> Result<()> {
 /// manipulation (no_std-clean); `sqrt` is gated behind `std`. Saturating float→int uses
 /// Rust's `as` cast, which matches wasm exactly (NaN→0, saturates to min/max).
 fn exec_float(frame: &mut Frame, op: Op) -> Result<()> {
-    match op as u8 {
+    match op as u16 {
         // f32 / f64 comparison (result i32) — IEEE ordering; NaN compares false (ne true).
         0x5b..=0x60 => {
             let b = frame.pop_f32();
